+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Added
  - ```depth``` field added to ```/markets``` endpoint.
 
+### Fixed
+ - Fixed onOrders query
+
 ## [1.0.2] - 2018-05-17
 
 ### Changed