+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [1.1.0] - 2018-05-27
 
-## Added
+### Added
  - ```depth``` field added to ```/markets``` endpoint.
 
 ### Fixed